Created an error message and added a race choice
Created the function errorMessage that gets called whenever the user does not enter one of the options that are listed. Also added races that the user can pick that will modify their current stats.

Also working on the choice that the user makes such as going into a tavern or into a forest.