Added roads and different locations in these areas
Added random stuff to these locations but I mostly worked on the right path, which introduces the player to a town called Verdana. This is a nod to the font Verdana, and CSS coders will probably get the reference. Furthermore, added a combat sequence between the players and the guards.