added a new race and a new variable
I added the race "asian" and I added a new variable "intelligence" which for asians is very high. This, of course, is an undocumented feature and will be added as a joke that only people who try to put asian in as a race or know the source code will be able to get. 

Hopefully this code change will add more humor to my source code.